move Presenter into its own file
