+++ CalculatorApp/CalculatorApp/PresentableResult.swift
@@ -1,0 +1,12 @@
+//
+//  PresentableResult.swift
+//  CalculatorApp
+//
+//  Created by Fatih Sağlam on 2.01.2024.
+//
+
+struct PresentableResult {
+    var totalAmount: String
+    var currentlyPaying: String
+    var remainingMonths: String
+}
